+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Create runtime subclass of SwaggerClient
 - Make default client and model type names more consistent
 - Use `typing.Mapping` instead of `typing.Dict` in generated stub file
+- Remove Black postprocessor
 
 ## 0.1.0